feat(web): ✨ update startup detail link
