+++ packages/web/src/pages/dashboard/components/ParticipatedCard.tsx
@@ -14,16 +14,16 @@ const StartupCard = defineComponent({
   setup(props, context) {
     const router = useRouter()
     const toStartDetail = () => {
-      router.push({ path: '/startup/detail', query: { startupId: props.startup.id } })
+      router.push({ path: `/startup/${props.startup.id}` })
     }
     return () => (
-      <div class="flex h-28 w-full items-center cursor-pointer" onClick={toStartDetail}>
+      <div class="cursor-pointer flex h-28 w-full items-center" onClick={toStartDetail}>
         <div class="flex h-full w-22 items-center">
           <UStartupLogo src={props.startup.logo} width="8" height="8" class="h-18 w-18" />
         </div>
         <div class="flex h-full border-b-1 ml-6 w-full items-center">
           <div class="content">
-            <div class="u-title1 mb-2">{props.startup.name}</div>
+            <div class="mb-2 u-title1">{props.startup.name}</div>
             <div class="divide-x">
               {props.startup.hashTags.map((tag, i) => {
                 return i + 1 < 4 ? (
