+++ packages/web/src/pages/startup/components/crowdfunding/CrowdfundingMiniCard.tsx
