+++ packages/web/src/pages/startup/components/team/index.tsx
